moving FullScreenOverlay to dedicated file
