+++ editor/src/components/github/github-clone-overlay.tsx
@@ -1,7 +1,3 @@
-/** @jsxRuntime classic */
-/** @jsx jsx */
-/** @jsxFrag React.Fragment */
-import { css, jsx, keyframes } from '@emotion/react'
 import React from 'react'
 import { GithubOperations } from '../../core/shared/github/operations'
 import { forceNotNull } from '../../core/shared/optional-utils'
@@ -9,9 +5,10 @@ import { NO_OP } from '../../core/shared/utils'
 import { totallyEmptyDefaultProject } from '../../sample-projects/sample-project-utils'
 import invariant from '../../third-party/remix/invariant'
 import { useOnClickAuthenticateWithGithub } from '../../utils/github-auth'
-import { Dialog, FormButton, UtopiaStyles, colorTheme } from '../../uuiui'
+import { Dialog, FormButton } from '../../uuiui'
 import type { EditorDispatch } from '../editor/action-types'
 import { setGithubState } from '../editor/actions/action-creators'
+import { FullScreenOverlay } from '../editor/full-screen-overlay'
 import { useDispatch } from '../editor/store/dispatch-context'
 import { type EditorStorePatched, type GithubRepoWithBranch } from '../editor/store/editor-state'
 import { Substores, useEditorState, useRefEditorState } from '../editor/store/store-hook'
@@ -164,78 +161,6 @@ const awaitLoadActionDispatchedByPersistenceMachine = (): Promise<{ projectId: s
     }
     PubSub.subscribe(LoadActionsDispatched, listener)
   })
-}
-
-const GithubCloneDialogBox = (props: React.PropsWithChildren<unknown>) => {
-  return (
-    <div
-      style={{
-        display: 'flex',
-        flexDirection: 'column',
-        width: 300,
-        opacity: 1,
-        fontSize: 12,
-        fontWeight: 500,
-        background: colorTheme.contextMenuBackground.value,
-        border: `1px solid ${colorTheme.neutralBorder.value}`,
-        padding: 30,
-        borderRadius: 2,
-        boxShadow: UtopiaStyles.shadowStyles.high.boxShadow,
-        whiteSpace: 'initial',
-      }}
-    >
-      {props.children}
-    </div>
-  )
-}
-
-function handleEventNoop(e: React.MouseEvent | React.KeyboardEvent) {
-  e.stopPropagation()
-  e.preventDefault()
-}
-
-export const FullScreenOverlay = (
-  props: React.PropsWithChildren<{ style?: React.CSSProperties }>,
-) => {
-  const anim = keyframes`
-    from {
-      opacity: 0;
-    }
-    to {
-      opacity: 0.2;
-    }
-  `
-
-  return (
-    <div
-      onMouseDown={handleEventNoop}
-      onMouseUp={handleEventNoop}
-      onClick={handleEventNoop}
-      onKeyDown={handleEventNoop}
-      onKeyUp={handleEventNoop}
-      style={{
-        position: 'fixed',
-        top: 0,
-        left: 0,
-        right: 0,
-        bottom: 0,
-        backgroundColor: '#00000033',
-        zIndex: 30,
-        transition: 'all .1s ease-in-out',
-        display: 'flex',
-        flexDirection: 'column',
-        alignItems: 'center',
-        justifyContent: 'center',
-        cursor: 'wait',
-        ...props.style,
-      }}
-      css={css`
-        animation: ${anim} 0.3s ease-in-out;
-      `}
-    >
-      {props.children}
-    </div>
-  )
 }
 
 export function getGithubRepoToLoad(urlSearchParams: string): GithubRepoWithBranch | null {
